cpp: darts - iteration 2

+++ cpp/darts/darts.h
@@ -2,7 +2,7 @@
 #define DARTS_H
 
 namespace darts {
-    int score(double x, double y);
+    int score(float x, float y);
 } // namespace darts
 
 #endif //DARTS_H
